+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorter/CommonClass.java
@@ -1,8 +1,0 @@
-package com.nemonotfound.nemos.inventory.sorter;
-
-public class CommonClass {
-
-    public static void init() {
-        Constants.LOG.info("Thank you for using Nemo's Inventory Sorter!");
-    }
-}