+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/CommonClass.java
@@ -1,0 +1,8 @@
+package com.nemonotfound.nemos.inventory.sorting;
+
+public class CommonClass {
+
+    public static void init() {
+        Constants.LOG.info("Thank you for using Nemo's Inventory Sorting!");
+    }
+}