+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/Constants.java
@@ -1,10 +1,10 @@
-package com.nemonotfound.nemos.inventory.sorter;
+package com.nemonotfound.nemos.inventory.sorting;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class Constants {
 
-	public static final String MOD_ID = "nemos_inventory_sorter";
+	public static final String MOD_ID = "nemos_inventory_sorting";
 	public static final Logger LOG = LoggerFactory.getLogger(MOD_ID);
 }