+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/client/gui/components/SortAlphabeticallyReversedButton.java
@@ -1,4 +1,4 @@
-package com.nemonotfound.nemos.inventory.sorter.client.gui.components;
+package com.nemonotfound.nemos.inventory.sorting.client.gui.components;
 
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.network.chat.Component;
@@ -8,7 +8,7 @@ import net.minecraft.world.item.ItemStack;
 import java.util.Comparator;
 import java.util.Map;
 
-import static com.nemonotfound.nemos.inventory.sorter.Constants.MOD_ID;
+import static com.nemonotfound.nemos.inventory.sorting.Constants.MOD_ID;
 
 public class SortAlphabeticallyReversedButton extends AbstractSortAlphabeticallyButton {