+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/mixin/ShulkerBoxScreenMixin.java
@@ -1,7 +1,7 @@
-package com.nemonotfound.nemos.inventory.sorter.mixin;
+package com.nemonotfound.nemos.inventory.sorting.mixin;
 
-import com.nemonotfound.nemos.inventory.sorter.client.gui.components.SortAlphabeticallyButton;
-import com.nemonotfound.nemos.inventory.sorter.client.gui.components.SortAlphabeticallyReversedButton;
+import com.nemonotfound.nemos.inventory.sorting.client.gui.components.SortAlphabeticallyButton;
+import com.nemonotfound.nemos.inventory.sorting.client.gui.components.SortAlphabeticallyReversedButton;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.gui.screens.inventory.ShulkerBoxScreen;
 import net.minecraft.network.chat.Component;
@@ -25,15 +25,15 @@ public abstract class ShulkerBoxScreenMixin extends AbstractContainerScreen<Ches
         int size = 12;
 
         //TODO: Change Component
-        SortAlphabeticallyButton sortAlphabeticallyButton = new SortAlphabeticallyButton(nemosInventorySorter$getLeftPosWithOffset(40), y, size, size, Component.literal("S"), this);
-        SortAlphabeticallyReversedButton sortAlphabeticallyReversedButton = new SortAlphabeticallyReversedButton(nemosInventorySorter$getLeftPosWithOffset(22), y, size, size, Component.literal("S"), this);
+        SortAlphabeticallyButton sortAlphabeticallyButton = new SortAlphabeticallyButton(nemosInventorySorting$getLeftPosWithOffset(40), y, size, size, Component.literal("S"), this);
+        SortAlphabeticallyReversedButton sortAlphabeticallyReversedButton = new SortAlphabeticallyReversedButton(nemosInventorySorting$getLeftPosWithOffset(22), y, size, size, Component.literal("S"), this);
 
         this.addRenderableWidget(sortAlphabeticallyButton);
         this.addRenderableWidget(sortAlphabeticallyReversedButton);
     }
 
     @Unique
-    private int nemosInventorySorter$getLeftPosWithOffset(int offset) {
+    private int nemosInventorySorting$getLeftPosWithOffset(int offset) {
         return this.leftPos + this.imageWidth - offset;
     }
 }