+++ neoforge/src/main/java/com/nemonotfound/nemos/inventory/sorting/NemosInventorySortingNeoForge.java
@@ -1,0 +1,13 @@
+package com.nemonotfound.nemos.inventory.sorting;
+
+
+import net.neoforged.fml.common.Mod;
+
+@Mod(Constants.MOD_ID)
+public class NemosInventorySortingNeoForge {
+
+    public NemosInventorySortingNeoForge() {
+        CommonClass.init();
+
+    }
+}